Refactor: Rename package from `recipe_shopping_list` to `recipeShoppingList`.
This commit updates the package name in `app/build.gradle.kts` and all relevant Kotlin files.
It also removes the XML declaration from `.idea/misc.xml`.
