+++ app/src/androidTest/java/com/example/recipeShoppingList/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.recipe_shopping_list
+package com.example.recipeShoppingList
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.recipe_shopping_list", appContext.packageName)
+        assertEquals("com.example.recipeShoppingList", appContext.packageName)
     }
 }
