+++ app/src/main/java/com/example/recipeShoppingList/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.recipe_shopping_list
+package com.example.recipeShoppingList
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.recipe_shopping_list.ui.theme.Recipe_Shopping_ListTheme
+import com.example.recipeShoppingList.ui.theme.Recipe_Shopping_ListTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
