+++ app/src/main/java/com/example/recipeShoppingList/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.example.recipe_shopping_list.ui.theme
+package com.example.recipeShoppingList.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
