+++ app/src/main/java/com/example/recipeShoppingList/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.recipe_shopping_list.ui.theme
+package com.example.recipeShoppingList.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
