+++ app/src/test/java/com/example/recipeShoppingList/domain/ShoppingList/AbstractShoppingListUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.recipe_shopping_list
+package com.example.recipeShoppingList
 
 import AbstractShoppingList
 import org.junit.Test
